fix: Change PlayerForm for visible team form.
