+++ adaptive_hockey_federation/main/forms.py
@@ -1,19 +1,23 @@
 import re
-from typing import Any
 
 from django import forms
 from django.core.exceptions import ValidationError
 from django.forms import (
     ModelChoiceField,
     ModelMultipleChoiceField,
-    MultipleChoiceField,
     Select,
     TextInput,
 )
-from django.shortcuts import get_object_or_404
 
 from core.constants import FORM_HELP_TEXTS, Role, StaffPosition
 from core.utils import max_date, min_date
+from main.fields import (
+    CityChoiceField,
+    CustomDiagnosisChoiceField,
+    CustomModelMultipleChoiceField,
+    CustomMultipleChoiceField,
+    StaffTeamMemberChoiceField,
+)
 from main.models import (
     City,
     Diagnosis,
@@ -25,69 +29,6 @@ from main.models import (
     Team,
 )
 from users.models import User
-
-
-class CustomMultipleChoiceField(MultipleChoiceField):
-    """Класс для расширения множественного поля выбора."""
-
-    def validate(self, value):
-        """Метод для валидации обязательного значения."""
-        if self.required and not value:
-            raise ValidationError(
-                self.error_messages["required"],
-                code="required",
-            )
-
-
-class CustomModelMultipleChoiceField(ModelMultipleChoiceField):
-    """Класс для расширения множественного поля выбора модели."""
-
-    def _check_values(self, value):
-        """Метод, проверяющий правильно ли указан список команд."""
-        try:
-            value = frozenset(value)
-        except TypeError:
-            raise ValidationError("Неверный список команд!")
-        value = list(value)
-        qs = Team.objects.filter(pk__in=value)
-        return qs
-
-
-class CustomDiagnosisChoiceField(ModelChoiceField):
-    """Самодельное поле для ввода диагноза."""
-
-    def __init__(self, label: str | None = None):
-        """
-        Метод инициализации экземпляра класса.
-
-        Добавляет виджет к полю выбора для поиска диагноза по названию.
-        """
-        super().__init__(
-            queryset=Diagnosis.objects.all(),
-            required=True,
-            widget=TextInput(
-                attrs={
-                    "list": "diagnosis",
-                    "placeholder": "Введите название диагноза",
-                },
-            ),
-            error_messages={
-                "required": "Пожалуйста, выберите диагноз из списка.",
-            },
-            label=label or "Выберите диагноз",
-        )
-
-    def clean(self, value: Any) -> Any:
-        """
-        Метод валидации поля.
-
-        Прежде, чем вызвать родительский метод, получает объект диагноза
-        (Diagnosis) по введенному названию, проверяет наличие введенного
-        наименования диагноза в БД.
-        """
-        if not value:
-            raise ValidationError(self.error_messages["required"])
-        return value
 
 
 class PlayerForm(forms.ModelForm):
@@ -103,6 +44,19 @@ class PlayerForm(forms.ModelForm):
     )
 
     diagnosis = CustomDiagnosisChoiceField()
+
+    team = CustomModelMultipleChoiceField(
+        queryset=None,
+        required=True,
+        help_text=FORM_HELP_TEXTS["player_teams"],
+        label="Команды",
+    )
+    available_teams = CustomModelMultipleChoiceField(
+        queryset=None,
+        required=False,
+        help_text=FORM_HELP_TEXTS["available_teams"],
+        label="Команды",
+    )
 
     class Meta:
         model = Player
@@ -159,15 +113,17 @@ class PlayerForm(forms.ModelForm):
     def __init__(self, *args, **kwargs):
         """Инициализация формы для игрока."""
         super().__init__(*args, **kwargs)
-        if self.instance.pk:
-            self.fields["nosology"].initial = self.instance.diagnosis.nosology
-            self.fields["diagnosis"].initial = self.instance.diagnosis
+        self.fields["team"].queryset = Team.objects.none()
+        self.fields[
+            "available_teams"
+        ].queryset = Team.objects.all().prefetch_related("city")
 
     def save(self, commit=True):
         """Метод создает и сохраняет объект игрока в базе данных."""
         instance = super().save(commit=False)
         if commit:
             instance.save()
+            instance.team.set(self.cleaned_data["team"])
         return instance
 
     def clean_identity_document(self):
@@ -207,23 +163,13 @@ class PlayerUpdateForm(PlayerForm):
         Расширяет team и available_teams кастомными полями выбора.
         """
         super(PlayerUpdateForm, self).__init__(*args, **kwargs)
-        queryset = self.instance.team.all()
-        self.fields["team"] = CustomModelMultipleChoiceField(
-            queryset=queryset,
-            required=True,
-            help_text=FORM_HELP_TEXTS["player_teams"],
-            label="Команды",
+        queryset = self.instance.team.all().prefetch_related("city")
+        self.fields["team"].queryset = queryset
+        self.fields["available_teams"].queryset = (
+            Team.objects.all().prefetch_related("city").difference(queryset)
         )
-        queryset_available = Team.objects.all().difference(queryset)
-        self.fields["available_teams"] = CustomModelMultipleChoiceField(
-            queryset=queryset_available,
-            required=False,
-            help_text=FORM_HELP_TEXTS["available_teams"],
-            label="Команды",
-        )
-        if self.instance.pk:
-            self.fields["nosology"].initial = self.instance.diagnosis.nosology
-            self.fields["diagnosis"].initial = self.instance.diagnosis
+        self.fields["nosology"].initial = self.instance.diagnosis.nosology
+        self.fields["diagnosis"].initial = self.instance.diagnosis
 
     def clean_number(self):
         """
@@ -264,101 +210,6 @@ class PlayerUpdateForm(PlayerForm):
             ).delete()
             instance.team.set(self.cleaned_data["team"])
         return instance
-
-
-class CityChoiceField(ModelChoiceField):
-    """Самодельное поле для выбора города."""
-
-    def __init__(self, label: str | None = None):
-        """
-        Метод инициализации экземпляра класса.
-
-        Добавляет виджет к полю выбора для поиска города по названию.
-        """
-        super().__init__(
-            queryset=City.objects.all(),
-            widget=TextInput(
-                attrs={
-                    "list": "cities",
-                    "placeholder": "Введите или выберите название города",
-                },
-            ),
-            required=True,
-            error_messages={
-                "required": "Пожалуйста, выберите город из списка.",
-            },
-            label=label or "Выберите город",
-        )
-
-    def clean(self, value: Any) -> Any:
-        """
-        Переопределенный метод родительского класса.
-
-        Прежде, чем вызвать родительский метод, получает объект города (
-        City) по введенному названию, проверяет наличие введенного
-        наименования города в БД. Если такого города в БД нет, то создает
-        соответствующий город (объект класса City) и возвращает его на
-        дальнейшую стандартную валидацию формы.
-        """
-        if not value:
-            raise ValidationError(self.error_messages["required"])
-
-        if value.isdigit():
-            return super().clean(value)
-        else:
-            city, created = City.objects.get_or_create(name=value)
-            return city
-
-
-class StaffTeamMemberChoiceField(ModelChoiceField):
-    """Самодельное поле выбора сотрудника команды."""
-
-    def __init__(self, team: Team, data_list: str, label: str | None = None):
-        """
-        Метод инициализации экземпляра класса.
-
-        Добавляет виджет к полю выбора для поиска сотрудника команды.
-        """
-        super().__init__(
-            queryset=StaffTeamMember.objects.all(),
-            widget=TextInput(
-                attrs={
-                    "list": data_list,
-                    "placeholder": "Начните ввод и выберите из списка",
-                },
-            ),
-            required=True,
-            error_messages={
-                "required": "Пожалуйста, выберите сотрудника из списка.",
-            },
-            label=label or "Выберите сотрудника",
-        )
-        self.team = team
-
-    def clean(self, value: Any) -> Any:
-        """
-        Переопределенный метод родительского класса.
-
-        Прежде, чем вызвать родительский метод, получает объект
-        StaffTeamMember и возвращает его на
-        дальнейшую стандартную валидацию формы.
-        """
-        if not value:
-            raise ValidationError(self.error_messages["required"])
-
-        if value.isdigit():
-            value = value
-        elif m := re.search(r"\d+\)\Z", value):
-            value = int(m.group()[:-1])
-        else:
-            raise ValidationError("Неверный формат введенных данных.")
-        staff_team_member = get_object_or_404(StaffTeamMember, id=value)
-        if StaffTeamMember.team.through.objects.filter(
-            staffteammember=staff_team_member,
-            team=self.team,
-        ).exists():
-            raise ValidationError("Этот сотрудник уже есть в команде.")
-        return super().clean(value)
 
 
 class TeamForm(forms.ModelForm):
